requirements.txt
fastapi
uvicorn
websockets
pymongo
python-dotenv
deepgram-sdk>=3.0.0
google-genai
python-multipart
pydantic
motor
requests
